rename: useGeolocation을 공통 hook 폴더로 이동
